+++ web_src/src/hooks/useCanvasData.ts
@@ -1797,6 +1797,25 @@ export const useUpdateCanvasConsole = (
 export type CanvasConsoleQueryResult = ReturnType<typeof useCanvasConsole>;
 export type UpdateCanvasConsoleMutationResult = ReturnType<typeof useUpdateCanvasConsole>;
 
+function buildRepositoryFileUrl(canvasId: string, path: string): string {
+  const params = new URLSearchParams({ path });
+  return `/api/v1/canvases/${encodeURIComponent(canvasId)}/repository/file?${params.toString()}`;
+}
+
+async function fetchRepositoryFileContent(canvasId: string, path: string): Promise<string> {
+  const response = await fetch(buildRepositoryFileUrl(canvasId, path), {
+    credentials: "include",
+    headers: withOrganizationHeader().headers,
+  });
+
+  if (!response.ok) {
+    const message = await response.text();
+    throw new Error(message || `Failed to load ${path}`);
+  }
+
+  return response.text();
+}
+
 export const useCanvasRepository = (canvasId: string, enabled: boolean = true) => {
   return useQuery({
     queryKey: canvasKeys.repository(canvasId),
@@ -1843,7 +1862,6 @@ export const useCanvasRepositoryFile = (
   return useQuery({
     queryKey: canvasKeys.repositoryFile(canvasId, normalizedPath, ref),
     queryFn: async () => {
-      const { fetchRepositoryFileContent } = await import("@/pages/app/files/lib/repository-files");
       const content = await fetchRepositoryFileContent(canvasId, normalizedPath);
       return {
         path: normalizedPath,
