+++ web_src/src/pages/app/files/lib/repository-files.ts
@@ -1,5 +1,3 @@
-import { withOrganizationHeader } from "@/lib/withOrganizationHeader";
-
 export function encodeRepositoryFileContent(content: string): string {
   const bytes = new TextEncoder().encode(content);
   let binary = "";
@@ -7,23 +5,4 @@ export function encodeRepositoryFileContent(content: string): string {
     binary += String.fromCharCode(...bytes.subarray(index, index + 0x8000));
   }
   return btoa(binary);
-}
-
-export function buildRepositoryFileUrl(canvasId: string, path: string): string {
-  const params = new URLSearchParams({ path });
-  return `/api/v1/canvases/${encodeURIComponent(canvasId)}/repository/file?${params.toString()}`;
-}
-
-export async function fetchRepositoryFileContent(canvasId: string, path: string): Promise<string> {
-  const response = await fetch(buildRepositoryFileUrl(canvasId, path), {
-    credentials: "include",
-    headers: withOrganizationHeader().headers,
-  });
-
-  if (!response.ok) {
-    const message = await response.text();
-    throw new Error(message || `Failed to load ${path}`);
-  }
-
-  return response.text();
 }
