+++ web_src/src/pages/app/index.tsx
@@ -92,8 +92,7 @@ import { useAppFiles } from "./files/useAppFiles";
 import { CanvasYamlModal } from "./CanvasYamlModal";
 import { useWorkflowViewSearchParams } from "./useWorkflowViewSearchParams";
 import { useFilesModeActions } from "./files/useFilesModeActions";
-import { resolveFilesHeaderVersionActions } from "./files/lib/resolve-files-header-version-actions";
-import { useFilesHeaderState } from "./files/useFilesHeaderState";
+import { resolveFilesHeaderVersionActions, useFilesHeaderState } from "./files/useFilesHeaderState";
 import { useMemoryModeActions } from "./useMemoryModeActions";
 import { useWorkflowHeaderEditActions } from "./useWorkflowHeaderEditActions";
 import { useWorkflowViewModeActions } from "./useWorkflowViewModeActions";
